% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schistox_functions.R
\name{administer_drug}
\alias{administer_drug}
\title{Administer drugs
function to administer drug to a specific variable (e.g. female_worms or eggs).
input the variable, the indices to apply to and the effectiveness of treatment
   administer_drug(humans, indices, drug_effectiveness)
administer mda drugs to chosen individuals in the population. If they adhere to the drugs, then they
reduce male and female worms with a given efficacy alongside removing eggs}
\usage{
administer_drug(humans, indices, drug_effectiveness)
}
\description{
Administer drugs
function to administer drug to a specific variable (e.g. female_worms or eggs).
input the variable, the indices to apply to and the effectiveness of treatment
   administer_drug(humans, indices, drug_effectiveness)
administer mda drugs to chosen individuals in the population. If they adhere to the drugs, then they
reduce male and female worms with a given efficacy alongside removing eggs
}
