% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schistox_functions.R
\name{administer_vaccine}
\alias{administer_vaccine}
\title{Administer vaccine ------------------------------------------------------
function to administer drug to a specific variable (e.g. female_worms or eggs).
input the variable, the indices to apply to and the effectiveness of treatment
   administer_vaccine(humans, indices, vaccine_effectiveness, vaccine_duration)
administer vaccine to chosen individuals in the population.
reduce male and female worms with a given efficacy alongside removing eggs and
adding to their vaccine status signifying that they will have increased immunity for a chosen period of time}
\usage{
administer_vaccine(humans, indices, vaccine_effectiveness, vaccine_duration)
}
\description{
Administer vaccine ------------------------------------------------------
function to administer drug to a specific variable (e.g. female_worms or eggs).
input the variable, the indices to apply to and the effectiveness of treatment
   administer_vaccine(humans, indices, vaccine_effectiveness, vaccine_duration)
administer vaccine to chosen individuals in the population.
reduce male and female worms with a given efficacy alongside removing eggs and
adding to their vaccine status signifying that they will have increased immunity for a chosen period of time
}
