% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schistox_functions.R
\name{birth_of_human}
\alias{birth_of_human}
\title{Birth of humans
add an individual to the population}
\usage{
birth_of_human(humans, pars)
}
\description{
Birth of humans
add an individual to the population
}
