% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schistox_functions.R
\name{calculate_worm_pairs}
\alias{calculate_worm_pairs}
\title{Calculate worm pairs in humans}
\usage{
calculate_worm_pairs(female_worms, male_worms)
}
\description{
function calculate_worm_pairs(humans)
    return min(sum(humans.female_worms), sum(humans.male_worms))
end
calculate how many pairs of worms there are in each human host
}
