% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schistox_functions.R
\name{cercariae_death}
\alias{cercariae_death}
\title{kill cercariae in the environment
function to kill cercariae in the environment
Kill a chosen proportion of cercariae in the environment governed by the
cercariae_survival parameter in the pars struct. This parameter governs what
proportion of cercariae survive for for one additional day, so if the time step
is greater than one, we have to calculate the correct proportion who die over
the chosen time step}
\usage{
cercariae_death(cercariae, miracidia, pars)
}
\description{
kill cercariae in the environment
function to kill cercariae in the environment
Kill a chosen proportion of cercariae in the environment governed by the
cercariae_survival parameter in the pars struct. This parameter governs what
proportion of cercariae survive for for one additional day, so if the time step
is greater than one, we have to calculate the correct proportion who die over
the chosen time step
}
