% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schistox_functions.R
\name{cercariae_uptake}
\alias{cercariae_uptake}
\title{Cercariae uptake --------------------------------------------------------
cercariae_uptake(humans, cercariae, miracidia, pars)
uptake cercariae into humans, whilst updating cercariae with miracidia.
Uptaken cercariae immediately become worms in this formulation}
\usage{
cercariae_uptake(humans, cercariae, miracidia, pars)
}
\description{
Cercariae uptake --------------------------------------------------------
cercariae_uptake(humans, cercariae, miracidia, pars)
uptake cercariae into humans, whilst updating cercariae with miracidia.
Uptaken cercariae immediately become worms in this formulation
}
