% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schistox_functions.R
\name{collect_prevs}
\alias{collect_prevs}
\title{Store prevalence and sac prevalence -------------------------------------
when we run multiple simulations, we store them in an array. This function will store the prevalence and sac prevalence
collect multiple prevalences within the population and store in appropriate arrays}
\usage{
collect_prevs(
  times,
  prev,
  sac_prev,
  high_burden,
  high_burden_sac,
  adult_prev,
  high_adult_burden,
  record,
  run
)
}
\description{
Store prevalence and sac prevalence -------------------------------------
when we run multiple simulations, we store them in an array. This function will store the prevalence and sac prevalence
collect multiple prevalences within the population and store in appropriate arrays
}
