% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schistox_functions.R
\name{create_contact_settings}
\alias{create_contact_settings}
\title{Contact settings --------------------------------------------------------
create the age specific contact settings given the scenario
This will create age dependent contact rates based on the scenario for simulation which is input. This is either
   "low adult", "moderate adult" or "high adult"}
\usage{
create_contact_settings(scenario)
}
\description{
Contact settings --------------------------------------------------------
create the age specific contact settings given the scenario
This will create age dependent contact rates based on the scenario for simulation which is input. This is either
   "low adult", "moderate adult" or "high adult"
}
