% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schistox_functions.R
\name{create_population}
\alias{create_population}
\title{Initial population ------------------------------------------------------
This will create the initial human population with randomly chosen age, and gender.
Predisposition is taken to be gamma distributed
There is also a male and female adjustment to predisposition adjusting for gender specific behaviour
In addition to this, it will create the initial miracidia environment vector}
\usage{
create_population(pars)
}
\description{
Initial population ------------------------------------------------------
This will create the initial human population with randomly chosen age, and gender.
Predisposition is taken to be gamma distributed
There is also a male and female adjustment to predisposition adjusting for gender specific behaviour
In addition to this, it will create the initial miracidia environment vector
}
