% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schistox_functions.R
\name{create_population_specified_ages}
\alias{create_population_specified_ages}
\title{Initial human population with an age distribution
This will create the initial human population with an age distribution
specified by the spec_ages variable
Predisposition is taken to be gamma distributed. There is also a male and female
adjustment to predisposition adjusting for gender specific behaviour
In addition to this, it will create the initial miracidia environment vector}
\usage{
create_population_specified_ages(pars)
}
\description{
Initial human population with an age distribution
This will create the initial human population with an age distribution
specified by the spec_ages variable
Predisposition is taken to be gamma distributed. There is also a male and female
adjustment to predisposition adjusting for gender specific behaviour
In addition to this, it will create the initial miracidia environment vector
}
