% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schistox_functions.R
\name{egg_production}
\alias{egg_production}
\title{Number of eggs produced -------------------------------------------------
function to calculate the number of eggs produced
 this is done by choosing from a negative binomial distribution for each worms,
 where the mean and aggregation parameters are calculated as in the
"Refined stratified-worm-burden models that incorporate specific biological features
of human and snail hosts provide better estimates of Schistosoma diagnosis,
transmission, and control" paper
for julia the negative binomial describes the number of failures before
 the given number of successes
 in a collection of independent Bernoulli trials.
 we need to specify a probability of success, and a given number of
 successes, which are derived
 from the mean and aggregation in the function below
inputs
r - aggregation factor for NB distribution
   egg_production!(humans, pars)
function to produce eggs for individuals, dependent on how many worms they have
       and the max fecundity and density dependent fecundity of the population}
\usage{
egg_production(humans, pars)
}
\description{
Number of eggs produced -------------------------------------------------
function to calculate the number of eggs produced
 this is done by choosing from a negative binomial distribution for each worms,
 where the mean and aggregation parameters are calculated as in the
"Refined stratified-worm-burden models that incorporate specific biological features
of human and snail hosts provide better estimates of Schistosoma diagnosis,
transmission, and control" paper
for julia the negative binomial describes the number of failures before
 the given number of successes
 in a collection of independent Bernoulli trials.
 we need to specify a probability of success, and a given number of
 successes, which are derived
 from the mean and aggregation in the function below
inputs
r - aggregation factor for NB distribution
   egg_production!(humans, pars)
function to produce eggs for individuals, dependent on how many worms they have
       and the max fecundity and density dependent fecundity of the population
}
