% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schistox_functions.R
\name{egg_production_increasing}
\alias{egg_production_increasing}
\title{egg production increasing
   egg_production_increasing!(humans, pars)
function to produce eggs for individuals, dependent on how many worms they have
       and the max fecundity and density dependent fecundity of the population}
\usage{
egg_production_increasing(humans, pars)
}
\description{
egg production increasing
   egg_production_increasing!(humans, pars)
function to produce eggs for individuals, dependent on how many worms they have
       and the max fecundity and density dependent fecundity of the population
}
