% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schistox_functions.R
\name{get_death_age}
\alias{get_death_age}
\title{Age for death of an individual ------------------------------------------
function to generate an age for death of an individual
This will create the initial human population with randomly chosen age, and gender.
Predisposition is taken to be gamma distributed
There is also a male and female adjustment to predisposition adjusting for gender specific behaviour
In addition to this, it will create the initial miracidia environment vector}
\usage{
get_death_age(pars)
}
\description{
Age for death of an individual ------------------------------------------
function to generate an age for death of an individual
This will create the initial human population with randomly chosen age, and gender.
Predisposition is taken to be gamma distributed
There is also a male and female adjustment to predisposition adjusting for gender specific behaviour
In addition to this, it will create the initial miracidia environment vector
}
