% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schistox_functions.R
\name{Human}
\alias{Human}
\title{Human -------------------------------------------------------------------
This struct contains the information about a human individual. This contains age, the pre determined age of death, community they are in,
   their gender, predisposition to picking up cercariae, the number of larvae, female and male worms and eggs in the individual along with
   a count of total lifetime eggs. Also it has their age dependent contact rate, adherence and access to interventions.}
\usage{
Human(
  age,
  death_age,
  gender,
  predisposition,
  female_worms,
  male_worms,
  eggs,
  vac_status,
  age_contact_rate,
  adherence,
  access,
  community,
  relative_contact_rate,
  uptake_rate,
  acquired_immunity,
  total_worms,
  larvae,
  last_uptake
)
}
\description{
Human -------------------------------------------------------------------
This struct contains the information about a human individual. This contains age, the pre determined age of death, community they are in,
   their gender, predisposition to picking up cercariae, the number of larvae, female and male worms and eggs in the individual along with
   a count of total lifetime eggs. Also it has their age dependent contact rate, adherence and access to interventions.
}
