% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schistox_functions.R
\name{human_larvae_maturity}
\alias{human_larvae_maturity}
\title{Human larvae maturity ---------------------------------------------------
This will mature the human larvae into worms after a chosen number of days, which is specified
by the human_larvae_maturity_time parameter in the pars struct}
\usage{
human_larvae_maturity(humans, pars)
}
\description{
Human larvae maturity ---------------------------------------------------
This will mature the human larvae into worms after a chosen number of days, which is specified
by the human_larvae_maturity_time parameter in the pars struct
}
