% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schistox_functions.R
\name{load_population_from_file}
\alias{load_population_from_file}
\title{load_population_from_file(filename)
load the environmental variables saved in the specified file}
\usage{
load_population_from_file(filename)
}
\description{
load_population_from_file(filename)
load the environmental variables saved in the specified file
}
