% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schistox_functions.R
\name{mda}
\alias{mda}
\title{Mass drug administration ------------------------------------------------
function for mass drug administration
currently there is no correlation between individuals chosen each time
 mda(humans, mda_coverage, min_age_mda, max_age_mda, mda_effectiveness, mda_gender)
administer mda in the population. This includes choosing individuals between specified ages,
having a certain level of coverage and taking access and adherence into consideration}
\usage{
mda(
  humans,
  mda_coverage,
  min_age_mda,
  max_age_mda,
  mda_effectiveness,
  mda_gender
)
}
\description{
Mass drug administration ------------------------------------------------
function for mass drug administration
currently there is no correlation between individuals chosen each time
 mda(humans, mda_coverage, min_age_mda, max_age_mda, mda_effectiveness, mda_gender)
administer mda in the population. This includes choosing individuals between specified ages,
having a certain level of coverage and taking access and adherence into consideration
}
