% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schistox_functions.R
\name{mda_information}
\alias{mda_information}
\title{mda_information ---------------------------------------------------------
This function contains the information for the mda, storing the coverage, minimum and maximum age targeted,
gender, drug efficacy and the time for the mda to be done}
\usage{
mda_information(
  mda_information,
  coverage,
  min_age,
  max_age,
  gender,
  effectiveness,
  time
)
}
\description{
mda_information ---------------------------------------------------------
This function contains the information for the mda, storing the coverage, minimum and maximum age targeted,
gender, drug efficacy and the time for the mda to be done
}
