% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schistox_functions.R
\name{out}
\alias{out}
\title{Out function -----------------------------------------------------------
This function contains the different outputs we are interested in recording. This is the
overall population burden, with categories for low, moderate and heavy burdens, along with
separate categories for the school age children and adults. Along with these, the time of each
result is recorded, so we can subsequently see the prevalence of the outbreak over time.}
\usage{
out(
  population_burden,
  sac_burden,
  adult_burden,
  pop_prev,
  sac_prev,
  adult_prev,
  sac_pop,
  adult_pop,
  final_ages,
  recorded_eggs,
  time
)
}
\description{
Out function -----------------------------------------------------------
This function contains the different outputs we are interested in recording. This is the
overall population burden, with categories for low, moderate and heavy burdens, along with
separate categories for the school age children and adults. Along with these, the time of each
result is recorded, so we can subsequently see the prevalence of the outbreak over time.
}
