% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schistox_functions.R
\name{run_repeated_sims_no_births_deaths}
\alias{run_repeated_sims_no_births_deaths}
\title{repeat simulations ------------------------------------------------------
repeat simulations where we allow mdas and vaccination, but keep the population the same by adding a birth for every death
run multiple simulations where aging of the population is not included and larvae are uptaken by humans as worms}
\usage{
run_repeated_sims_no_births_deaths(
  filename,
  num_time_steps,
  mda_info,
  vaccine_info,
  num_repeats
)
}
\description{
repeat simulations ------------------------------------------------------
repeat simulations where we allow mdas and vaccination, but keep the population the same by adding a birth for every death
run multiple simulations where aging of the population is not included and larvae are uptaken by humans as worms
}
