% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schistox_functions.R
\name{run_repeated_sims_no_births_deaths_human_larvae}
\alias{run_repeated_sims_no_births_deaths_human_larvae}
\title{repeat simulations
run multiple simulations where aging of the population is not included and larvae are uptaken by humans as larvae}
\usage{
run_repeated_sims_no_births_deaths_human_larvae(
  filename,
  num_time_steps,
  mda_info,
  vaccine_info,
  num_repeats
)
}
\description{
repeat simulations
run multiple simulations where aging of the population is not included and larvae are uptaken by humans as larvae
}
