% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schistox_functions.R
\name{run_repeated_sims_no_births_deaths_increasing}
\alias{run_repeated_sims_no_births_deaths_increasing}
\title{repeat simulations
repeat simulations where we allow mdas and vaccination, but keep the population the same by adding a birth for every death
run multiple simulations where aging of the population is not included and larvae are uptaken by humans as worms, and egg production is
monotonically increasing}
\usage{
run_repeated_sims_no_births_deaths_increasing(
  filename,
  num_time_steps,
  mda_info,
  vaccine_info,
  num_repeats
)
}
\description{
repeat simulations
repeat simulations where we allow mdas and vaccination, but keep the population the same by adding a birth for every death
run multiple simulations where aging of the population is not included and larvae are uptaken by humans as worms, and egg production is
monotonically increasing
}
