% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schistox_functions.R
\name{run_repeated_sims_no_population_change_human_larvae}
\alias{run_repeated_sims_no_population_change_human_larvae}
\title{repeat simulations
repeat simulations where we allow mdas and vaccination, but keep the population the same by adding a birth for every death
run multiple simulations where the population is aged, but each death is replaced by a birth and larvae are uptaken by humans as larvae}
\usage{
run_repeated_sims_no_population_change_human_larvae(
  filename,
  num_time_steps,
  mda_info,
  vaccine_info,
  num_repeats
)
}
\description{
repeat simulations
repeat simulations where we allow mdas and vaccination, but keep the population the same by adding a birth for every death
run multiple simulations where the population is aged, but each death is replaced by a birth and larvae are uptaken by humans as larvae
}
