% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schistox_functions.R
\name{run_repeated_sims_no_population_change_increasing}
\alias{run_repeated_sims_no_population_change_increasing}
\title{repeated simulations ----------------------------------------------------
   run_repeated_sims_no_population_change(filename, num_time_steps, mda_info, vaccine_info, num_repeats)
run multiple simulations where aging of the population is not included, larvae are uptaken by humans as worms}
\usage{
run_repeated_sims_no_population_change_increasing(
  filename,
  num_time_steps,
  mda_info,
  vaccine_info,
  num_repeats
)
}
\description{
repeated simulations ----------------------------------------------------
   run_repeated_sims_no_population_change(filename, num_time_steps, mda_info, vaccine_info, num_repeats)
run multiple simulations where aging of the population is not included, larvae are uptaken by humans as worms
}
