% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schistox_functions.R
\name{specified_age_distribution}
\alias{specified_age_distribution}
\title{construct the set of ages -----------------------------------------------
function to construct the set of ages, with size N
specified_age_distribution(pars)
output ages according to a specified age distribution}
\usage{
specified_age_distribution(pars)
}
\description{
construct the set of ages -----------------------------------------------
function to construct the set of ages, with size N
specified_age_distribution(pars)
output ages according to a specified age distribution
}
