% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schistox_functions.R
\name{update_contact_rate}
\alias{update_contact_rate}
\title{update contact rates ----------------------------------------------------
function to update the contact rate of individuals in the population. This is necessary
as over time when people age, they will move through different age groups which have
different contact rates}
\usage{
update_contact_rate(humans, pars)
}
\description{
update contact rates ----------------------------------------------------
function to update the contact rate of individuals in the population. This is necessary
as over time when people age, they will move through different age groups which have
different contact rates
}
