% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schistox_functions.R
\name{update_env_constant_population_human_larvae}
\alias{update_env_constant_population_human_larvae}
\title{update env_sonstant population ------------------------------------------
   update_env_constant_population_human_larvae(num_time_steps, humans,  miracidia, cercariae, pars, mda_info, vaccine_info)
update the population for a given length of time. Here we include deaths and for each death an individual is immediately born.
Interventions are included in this function and larvae are uptaken as larvae in the humans.}
\usage{
update_env_constant_population_human_larvae(
  num_time_steps,
  humans,
  miracidia,
  cercariae,
  pars,
  mda_info,
  vaccine_info
)
}
\description{
update env_sonstant population ------------------------------------------
   update_env_constant_population_human_larvae(num_time_steps, humans,  miracidia, cercariae, pars, mda_info, vaccine_info)
update the population for a given length of time. Here we include deaths and for each death an individual is immediately born.
Interventions are included in this function and larvae are uptaken as larvae in the humans.
}
