% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schistox_functions.R
\name{update_env_no_births_deaths_increasing}
\alias{update_env_no_births_deaths_increasing}
\title{update env no. births and deaths increasing
   update_env_no_births_deaths_increasing(num_time_steps, humans,  miracidia, cercariae, pars, mda_info, vaccine_info)
update the population for a given length of time. Here we do not include births or deaths and individuals do not age
Interventions are included in this function and larvae are uptaken as immediately as worms and egg production is monotonically increasing}
\usage{
update_env_no_births_deaths_increasing(
  num_time_steps,
  humans,
  miracidia,
  cercariae,
  pars,
  mda_info,
  vaccine_info
)
}
\description{
update env no. births and deaths increasing
   update_env_no_births_deaths_increasing(num_time_steps, humans,  miracidia, cercariae, pars, mda_info, vaccine_info)
update the population for a given length of time. Here we do not include births or deaths and individuals do not age
Interventions are included in this function and larvae are uptaken as immediately as worms and egg production is monotonically increasing
}
