% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schistox_functions.R
\name{update_env_to_equilibrium_human_larvae}
\alias{update_env_to_equilibrium_human_larvae}
\title{update the pop for a given length of time
update_env_to_equilibrium_human_larvae(num_time_steps, humans, miracidia, cercariae, pars)
update the population for a given length of time. Here we do not age the population or include birth, deaths or interventions and for this
function larvae are uptaken from the environment into a larvae category in the humans, rather than immediately becoming worms}
\usage{
update_env_to_equilibrium_human_larvae(
  num_time_steps,
  humans,
  miracidia,
  cercariae,
  pars
)
}
\description{
update the pop for a given length of time
update_env_to_equilibrium_human_larvae(num_time_steps, humans, miracidia, cercariae, pars)
update the population for a given length of time. Here we do not age the population or include birth, deaths or interventions and for this
function larvae are uptaken from the environment into a larvae category in the humans, rather than immediately becoming worms
}
