% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schistox_functions.R
\name{vac_decay}
\alias{vac_decay}
\title{add vaccination to population -------------------------------------------
function to add vaccination to population
update vaccine information ----------------------------------------------
function to update vaccine information =#
vaccine decay -----------------------------------------------------------
decrease vaccination status for each person by 1 each day}
\usage{
vac_decay(humans, pars)
}
\description{
add vaccination to population -------------------------------------------
function to add vaccination to population
update vaccine information ----------------------------------------------
function to update vaccine information =#
vaccine decay -----------------------------------------------------------
decrease vaccination status for each person by 1 each day
}
