% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schistox_functions.R
\name{worm_maturity}
\alias{worm_maturity}
\title{Kill worms within human host --------------------------------------------
Worms die have a specified mean life span, and hence a rate of deaths per day .
the number of deaths, or maturing from each stage is dependent
on the number of worm stages and rate of aging through stages and dying.
This p is multiplied by the time scale of the simulation, so if 2 days
pass between consecutive time points, twice as many worms age and die
for human in humans
    println(human.eggs)
end
function to kill worms within human hosts, and if there is more than one stage for worm life,
to update how many worms are in each stage}
\usage{
worm_maturity(humans, pars)
}
\description{
Kill worms within human host --------------------------------------------
Worms die have a specified mean life span, and hence a rate of deaths per day .
the number of deaths, or maturing from each stage is dependent
on the number of worm stages and rate of aging through stages and dying.
This p is multiplied by the time scale of the simulation, so if 2 days
pass between consecutive time points, twice as many worms age and die
for human in humans
    println(human.eggs)
end
function to kill worms within human hosts, and if there is more than one stage for worm life,
to update how many worms are in each stage
}
