% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schistox_functions.R
\name{create_mda}
\alias{create_mda}
\title{create MDA --------------------------------------------------------------
function to create a set of mda's which will be performed regularly
first_mda_time specifies when this will first occur in years,
last_mda_time is the final mda in this block
regularity is how often to perform the mda in years.
specify the proportion of pre SAC, SAC and adults at each of these time points
also specify genders for these differect age groups, along with the effectiveness of mda}
\usage{
create_mda(
  pre_SAC_prop,
  SAC_prop,
  adult_prop,
  first_mda_time,
  last_mda_time,
  regularity,
  pre_SAC_gender,
  SAC_gender,
  adult_gender,
  mda_effectiveness
)
}
\description{
create MDA --------------------------------------------------------------
function to create a set of mda's which will be performed regularly
first_mda_time specifies when this will first occur in years,
last_mda_time is the final mda in this block
regularity is how often to perform the mda in years.
specify the proportion of pre SAC, SAC and adults at each of these time points
also specify genders for these differect age groups, along with the effectiveness of mda
}
