% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schistox_functions.R
\name{make_age_contact_rate_array}
\alias{make_age_contact_rate_array}
\title{Age dependent contact rate
This is a function to get age dependent contact rate.
the contact rates are taken from the
"What is required in terms of mass drug administration to interrupt the transmission
    of schistosome parasites in regions of endemic infection?" paper
at some point we may change this to be an input from a file instead}
\usage{
make_age_contact_rate_array(pars, scenario, input_ages, input_contact_rates)
}
\description{
Age dependent contact rate
This is a function to get age dependent contact rate.
the contact rates are taken from the
"What is required in terms of mass drug administration to interrupt the transmission
    of schistosome parasites in regions of endemic infection?" paper
at some point we may change this to be an input from a file instead
}
