% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schistox_functions.R
\name{Parameters}
\alias{Parameters}
\title{Parameters}
\usage{
Parameters(
  N = 500,
  time_step = 20,
  N_communities = 1,
  community_probs = 1,
  community_contact_rate = 1,
  density_dependent_fecundity = 7e-04,
  average_worm_lifespan = 5.7,
  max_age = 100,
  initial_worms = 10,
  initial_miracidia = 1e+08,
  initial_miracidia_days = 1,
  init_env_cercariae = 1e+08,
  worm_stages = 1,
  contact_rate = 0.18,
  max_fec_contact_rate_product = 0.8,
  max_fecundity = 20,
  age_contact_rates = input_contact_rates/sum(input_contact_rates),
  ages_for_contacts = c(4, 9, 15, 100),
  contact_rate_by_age_array = rep(0, times = max_age + 1),
  mda_adherence = 1,
  mda_access = 1,
  female_factor = 1,
  male_factor = 1,
  miracidia_maturity = 24,
  birth_rate = 28 * time_step/(1000 * 365),
  human_cercariae_prop = 1,
  predis_aggregation = 0.24,
  cercariae_survival = 0.05,
  miracidia_survival = 0.05,
  death_prob_by_age = c(0.0656, 0.0093, 0.003, 0.0023, 0.0027, 0.0038, 0.0044, 0.0048,
    0.0053, 0.0065, 0.0088, 0.0106, 0.0144, 0.021, 0.0333, 0.0529, 0.0851, 0.1366,
    0.2183, 0.2998, 0.3698, 1),
  ages_for_death = c(1, 5, 10, 15, 20, 25, 30, 35, 40, 45, 50, 55, 60, 65, 70, 75, 80,
    85, 90, 95, 100, 110),
  r = 0.03,
  vaccine_effectiveness = 0.86,
  drug_effectiveness = 0.86,
  spec_ages = c(7639, 7082, 6524, 5674, 4725, 4147, 3928, 3362, 2636, 1970, 1468, 1166,
    943, 718, 455, 244),
  ages_per_index = 5,
  record_frequency = 1/24,
  use_kato_katz = 0,
  kato_katz_par = 0.87,
  heavy_burden_threshold = 50,
  rate_acquired_immunity = 0,
  M0 = 20,
  human_larvae_maturity_time = 30,
  egg_sample_size = 1/100,
  input_ages = c(4, 9, 15, 100),
  input_contact_rates = c(0.01, 1.2, 1, 0.02),
  scenario = "high adult"
)
}
\arguments{
\item{N}{human population size}

\item{time_step}{length of time step (in days)}

\item{N_communities}{number of communities in the population sharing the same environmental source}

\item{community_probs}{probability of being in each community}

\item{community_contact_rate}{contact rate with the environment for each of the commununity}

\item{density_dependent_fecundity}{decrease in egg production per worm due to high density of worms}

\item{average_worm_lifespan}{average expectancy of a worm}

\item{max_age}{maximum age of individual}

\item{initial_worms}{initial no. of worms}

\item{initial_miracidia}{initial no. of miracidia in the environment}

\item{initial_miracidia_days}{no.of days miracidia will age into cercariae larvae}

\item{init_env_cercariae}{initial no of cercaria in the environment}

\item{worm_stages}{number of stages in the worm. Having 1 stage will result to a Gamma distribution}

\item{contact_rate}{global contact rate for the uptake of larvae from the environment}

\item{max_fec_contact_rate_product}{product of max fecundity and the contact rate in the population. Setting this to a desired value is often a good way to ensure that the epidemic stays within a reasonable range, as when the max fecundity increases, if the contact rate doesn't decrease appropriately, then the behaviour of the outbreak can be unrealistically difficult to control.}

\item{max_fecundity}{expected no. of eggs from a single worm}

\item{age_contact_rates}{contact rate for the uptake of larvae from the environment for the chosen age groups}

\item{ages_for_contacts}{age groups for specifying contact rates}

\item{contact_rate_by_age_array}{<- rep(0,times=max_age+1) array holding contact rate for each age}

\item{mda_adherence}{proportion of people who adhere to the MDA}

\item{mda_access}{proportion of people who have access to the MDA}

\item{female_factor}{factor for altering the contact rate for females, if we choose to have gender-specific behavior which affects contact rate}

\item{male_factor}{factor for altering the contact rate for males, if we choose to have gender-specific behavior which affects contact rate}

\item{miracidia_maturity}{no of days after which miracidias will mature to cercariae}

\item{birth_rate}{rate of birth of humans}

\item{human_cercariae_prop}{proportion of cercariae which are able to infect humans}

\item{predis_aggregation}{aggregation parameter for Poisson distributed egg production}

\item{cercariae_survival}{proportion of cercariae that survive from one time point to the next}

\item{miracidia_survival}{proportion of miracidia that survive from one time point to the next}

\item{death_prob_by_age}{probability of dying each year, specified by age}

\item{ages_for_death}{age ranges for death probailities}

\item{r}{aggregation parameter for negative binomially distributed egg production}

\item{vaccine_effectiveness}{efficacy of a vaccine if one is used}

\item{drug_effectiveness}{efficacy of a drug given during MDA}

\item{spec_ages}{number of individuals by age group}

\item{ages_per_index}{how many different ages we include in the spec_ages parameter}

\item{record_frequency}{how often we should record the prevalence in the population dusing simulation}

\item{use_kato_katz}{if 1, use Kato-Katz for egg count, if 0, do not use KK}

\item{kato_katz_par}{parameter for Gamma distribution if KK is used}

\item{heavy_burden_threshold}{number of eggs at which an individual is said to have a heavy infection}

\item{rate_acquired_immunity}{rate at which immunity will be acquired for individuals. This will be multiplied by the cumulative nymber of worms people have had during their life to decide the level of immunity acquired}

\item{M0}{if a particular formula of egg production is used, this parameter is required and is a proxy for mean worm burden}

\item{human_larvae_maturity_time}{length of time (in days) after which a cercariae uptake by a human will mature into a worm}

\item{egg_sample_size}{the proportion of eggs which are sampled from each individual every time we check their burden (between 0 and 1). 1= all eggs in the person are sampled. Typical value fpr a urine sample may be ~1/100}

\item{input_ages}{input ages for constructing contact array}

\item{input_contact_rates}{input contact rates}

\item{scenario}{can be one of "low adult", "moderate adult" or high adult"}
}
\description{
This function takes in all the parameters for the model
}
